Create iterator loop for yaml schema
Now need to break out the directory path logic into a recursive function
so it can handle subdirectories and even files later.

Then we can have a loop that says for every item in the directories
field generate paths. Then it recursively generates paths and then
creates the directories when done. As a result that function is gonna
need a reference to the base path, which should just be ./ but you never
know